ats-resume-matcher/requirements.txt

fastapi
uvicorn[standard]
pdfplumber
python-docx
scikit-learn
pydantic
python-multipart
streamlit
